Handle high-contrast checkbox columns
